Generate final snapshot identifiers (and better instance identifiers)
When spinning up a workspace instance, include the database engine and 8
chars from the SHA256 hash of the Instance Identifier of the source
instance.

Pre-generate the final snapshot identifier from the source instance
identifier and the timestamp.

Note that instance identifiers are limited to 63 chars whereas snapshot
identifiers are limited to 255.