Remove some TODOs
Explicit error checking in the locations indicated by the TODOs is
probably not necessary, as boto3 will raise an exception if anything
goes wrong.

If an exception is recoverable we can trap and deal with it locally, but
that will not be the case most of the time and we'll want to trap it
higher up and issue a CRITICAL error.