Optionally skip final snapshot creation
We don't want to leave our workspace instances hanging around; we always
want to clean them up. But if we create a final snapshot and share it
when one of the scrub tasks failed, we could be inadvertently sharing
sensitive data that hasn't been scrubbed. Add a parameter to the `cleanup`
function to control final snapshot creation.