Specify psycopg2 dependency with --no-binary
This one was a bit confusing. Installing psycopg2 (2.7) pulls down a
binary wheel for the current platform by default. However, this binary
wheel has a deprecation warning that pops up when you use the package.

The two upgrade paths that avoid the deprecation warning are:
 * Continue to use the binary distribution, but requesting it explicitly
   by installing the psycopg2-binary package instead (this isn't
   recommended for our use case)
 * Force installation from source, in which case it will be compiled
   locally against libpq.

The second one is what we're doing here.