Make the workspace instance name consistent and predictable
Drop the timestamps from the workspaces' instance identifiers.

The instance identifier is used to select the snapshots on the import
side, and having a timestamp in there adds complexity.

Continue to use a hash of the source instance identifier; this
guarantees we don't overflow the (relatively short) instance identifier
size limit of 63 chars.